Collapsing columns is now optional
Previously the column labels would always collapse into the header row with the index label. Consequently, the name of the column level would never be shown.

Now an attribute "collapse-columns" has been added. If this is set then the columns will collapse into the index name row of the header. If not then the column labels will go into their own row within the header.