Add formatting editing
In tab format now a format table is shown containing all columns, dtypes and formats. Clicking on the edit button opens a dialog that allows the user to edit the formatting options.

The options in the dialog get created dynamically from the format-specs. Fieldsets can be dependent on certain selection states.