Move smart defaults into viewer
By setting smart defaults (this is a calculated value depending on the table spec used if no explicit setting was provided) in the viewer we can also ensure that the settings popup has the actual setting as it is applied (even if it was calculated).

Note that smart defaults only work on first data load. If different sources are loaded into the same data-viewer then subsequent sources would just use whatever smart default was set by the first source.